Cda delete button working 100% of the time

+++ src/components/JournalComponent/index.js
@@ -14,9 +14,13 @@ export default function JournalComponent({ id, title, date, body, editClick, ...
   const [isOpen, setIsOpen] = useState(false);
 
   const deleteClick = id => {
+    const deletedId = id;
     API.deleteEntry(id).then(res => {
-      deleteReset();
-      setIsOpen(!isOpen)
+      if(deletedId !== res){
+        deleteReset();
+        setIsOpen(!isOpen)
+      }
+
     });
   };
